+++ tests/integration/go/test_go.py
@@ -6,10 +6,10 @@
 import logging
 import typing
 
-import pytest
 import requests
 from juju.application import Application
 from juju.model import Model
+from pytest_operator.plugin import OpsTest
 
 logger = logging.getLogger(__name__)
 WORKLOAD_PORT = 8080
@@ -101,3 +101,62 @@ async def test_prometheus_integration(
         else:
             logger.error("Application not scraped. Scraped targets: %s", active_targets)
             assert False, "Scrape Target not configured correctly"
+
+
+async def test_open_ports(
+    ops_test: OpsTest,
+    model: Model,
+    go_app: Application,
+    traefik_app: Application,
+    get_unit_ips,
+    external_hostname,
+):
+    """
+    arrange: after Go charm has been deployed.
+    act: integrate with the traefik charm with the ingress integration and change app-port configuration.
+    assert: charm opened ports should change accordingly.
+    """
+    await model.add_relation(traefik_app.name, go_app.name)
+    await model.wait_for_idle(apps=[go_app.name, traefik_app.name], status="active")
+    traefik_ip = (await get_unit_ips(traefik_app.name))[0]
+
+    juju_cmd = ["exec", "--unit", f"{go_app.name}/0", "opened-ports"]
+    _, opened_ports, _ = await ops_test.juju(*juju_cmd)
+    assert opened_ports.strip() == f"{WORKLOAD_PORT}/tcp"
+    assert (
+        requests.get(
+            f"http://{traefik_ip}",
+            headers={"Host": f"{ops_test.model_name}-{go_app.name}.{external_hostname}"},
+            timeout=5,
+        ).status_code
+        == 200
+    )
+
+    new_port = WORKLOAD_PORT + 10
+    await go_app.set_config({"app-port": str(new_port)})
+    await model.wait_for_idle(apps=[go_app.name, traefik_app.name], status="active")
+
+    _, opened_ports, _ = await ops_test.juju(*juju_cmd)
+    assert opened_ports.strip() == f"{new_port}/tcp"
+    assert (
+        requests.get(
+            f"http://{traefik_ip}",
+            headers={"Host": f"{ops_test.model_name}-{go_app.name}.{external_hostname}"},
+            timeout=5,
+        ).status_code
+        == 200
+    )
+
+    await go_app.set_config({"app-port": str(WORKLOAD_PORT)})
+    await model.wait_for_idle(apps=[go_app.name, traefik_app.name], status="active")
+
+    _, opened_ports, _ = await ops_test.juju(*juju_cmd)
+    assert opened_ports.strip() == f"{WORKLOAD_PORT}/tcp"
+    assert (
+        requests.get(
+            f"http://{traefik_ip}",
+            headers={"Host": f"{ops_test.model_name}-{go_app.name}.{external_hostname}"},
+            timeout=5,
+        ).status_code
+        == 200
+    )